backend-old/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
pydantic>=1.10.0